Rework tests
