+++ tox.ini
@@ -1,46 +1,28 @@
 [tox]
-;envlist = py27,py37,py38
-envlist = pylint,py37,py38
+envlist = py37,py38,py39,py310,pylint,docs
+skipsdist = True
+skip_missing_interpreters = True
 
-deps = -r{toxinidir}/requirements.txt
-       -r{toxinidir}/test-requirements.txt
-
+[base]
+deps =
+	-r requirements.txt
+	-r test-requirements.txt
+module = flaat
 
 [testenv]
-# install pytest in the virtualenv where commands will be executed
-commands =
-    # NOTE: you can run any command line tool here - not just tests
-    pip install -e .
-    python tests/tokentools_test.py
+deps = {[base]deps}
+commands = pytest
 
-[testenv:pep8]
-commands =
-    nosetests
+[testenv:py310]
+deps = {[base]deps}
+commands = pytest --cov={[base]module} --cov-report=term-missing
 
 [testenv:pylint]
-deps = pylint
-commands = 
-    pylint -E flaat/flaat_exceptions.py
-    pylint -E flaat/flaat_userinfo.py
-    pylint -E flaat/__init__.py
-    pylint -E flaat/issuertools.py
-    pylint -E flaat/tokentools.py
-
-
-[flake8]
-# H803 skipped on purpose per list discussion.
-# E123, E125 skipped as they are invalid PEP-8.
-
-show-source = True
-ignore = E123,E125,H803,H405
-builtins = _
-exclude=.venv,.git,.tox,dist,doc,*lib/python*,*egg,build
+deps = {[base]deps}
+commands = pylint --exit-zero {[base]module}
 
 [testenv:docs]
 deps =
-	sphinx
-	sphinx-argparse
-    -r{toxinidir}/requirements.txt
-commands =
-	sphinx-build "{toxinidir}/doc/source" "{toxinidir}/doc/build"
-	python -c 'import pathlib; print("Documentation available under file://\{0\}".format(pathlib.Path(r"{toxinidir}") / "doc" / "build" / "index.html"))'
+	-r requirements.txt
+	-r doc/source/requirements.txt
+commands = sphinx-build doc/source doc/build
